notification/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
httpx==0.25.2
python-dotenv==1.0.0
aiosqlite==0.19.0
tweepy==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3  # For XML parsing 